Made the solver algorithm and changed the generator
Generator changes:
I made it so that there is a simpler visualizations for the generated maze by turning every wall into a 1.
The generated maze now sits on another file, so that it can be used by the solving program

------------

Solver:
I realized that when trying to solve bigger mazes DFS (Depth First Search) is better than BFS (Breadth First Search). However, BFS is better in solving smaller mazes. So I made it work with both.
I also made a visual representation of the maze and the solution using the Pillow library.