Made an app that takes everything and puts it together
For the GUI I'm using Tkinter. The user can put the input in the width and height fields. For the making of the app I had to comment the usage of the generator.py and solver.py, so if you want to use just the generator script for example, you'll have to turn them into normal code